requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==3.0.1
attrs==22.1.0
chardet==4.0.0
charset-normalizer==2.1.1
discord.py==1.7.3
Django==4.1
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
yarl==1.8.1
